rp2 machine.PWM
micropython pull #/9e1b25a

docs/library/machine: Specify initial machine.PWM class.

This adds an initial specification of the machine.PWM class, to provide
a
way to generate PWM output that is portable across the different ports.
Such functionality may already be available in one way or another (eg
through a Timer object), but because configuring PWM via a Timer is very
port-specific, and because it's a common thing to do, it's beneficial to
have a top-level construct for it.

The specification in this commit aims to provide core functionality in a
minimal way.  It also somewhat matches most existing ad-hoc
implementations
of machine.PWM